Update mood requirement for commit subject/title
What:
Removes the requirement to use the imperative mood in the subject/title and instead leaves it up to the committer how to specify what the commit does/will do if the commit is applied.

Why:
People may not be familiar with what the imperative mood is and we want to allow committers to use what they prefer as long as it fits with our general style.