Add vision and guiding principles document
What: An initial version of the vision/mission for the project, and the principles/values that guide it.

Why: To communicate the spirit of the project and the direction it's going, and so we have something concrete that we can point to/update as necessary.

Notes: Initial inspiration and input taken from https://discord.com/channels/1210620048297955448/1212947061755088959 in the V-Formation Village Discord server as well as from conversations with the community.

Other sources of inspiration:
* https://www.linkedin.com/pulse/meow-wolf-announces-layoffs-why-i-quit-december-joanna-garner-sm0zc (first posted here: https://discord.com/channels/1210620048297955448/1211677680194289716/1237767024311144479)
* https://organiccouncil.ca/regenerative-organic-certification/
* https://principles.xrguild.org/ (first of several pages)
* https://phys.org/news/2024-08-drought-sicily-rainwater-dumped-sea.html
* https://www.youtube.com/watch?v=ccrmlHTUB2Q

V-Formation Village Discord server invite:
https://discord.gg/XD7jdR33Ba