Suggest Hubs relevant rule and mod guidelines
What: Suggests the addition of a Hubs-relevant rule
for the Hubs Discord original or first posts.
Includes examples of what does and does not follow the rule.
Also includes moderator guidelines on channels where
enforcement should not be as strict.

Why: We occasionally see metaverse, crypto, other
platform posts, or posts where any connection to
Hubs seems missing. Making this a server rule that
all members will help enforcement because especially
new members will have just agreed to it upon joining.
Therefore, members can be held to this rule. Moderators can,
if they have to, remove posts or at least minimally
ask the original poster to edit/add Hubs relevance.
Moderators should check out links, and/or investigate
the posting member, and remove posts if they are
not relevant to the Hubs community.

Note: If discussion approves adding this as a Discord rule,
 that will need to be done at Discord separately.

There are two key changes included in this rule:

1. The post must be obviously Hubs related OR the posting
member must make it relevant by adding context to their post
that makes the perceived connection between Hubs and the content clear.

2. It places some responsibility on the Moderators
to keep the rest of the community safe by checking links,
which might be spam, or by investigating the posting members
 by using publicly available data.