Add AI usage policy
What: Adds a new document containing rules for AI tool usage and guidelines on the disclosure of AI tool usage.

Why:
* To minimize legal risks to the project and help ensure AI assisted submissions are high quality, and can be handled smoothly and consistently.
* Pull requests that use AI tools have been increasing.

Notes:
This is intended to apply to all AI usage, whether it's writing/documentation, media, code, etc..  However, there may need to be area specific modifications or separate area specific policies may need to be broken off.

This has been the topic of heavy debate/discussion in Discord recently, so more modifications to this will likely follow, especially as court cases regarding AI are resolved.

This has been largely based off of the proposal at the end of Imaginer's initial AI concerns post in Discord and the disclosure guidelines suggested in a subsequent post by Imaginer, along with input from Doug Reeder at the Dev Meetups.

This doesn't address any form of detection of AI generated material, other than asking people, or the implementation of any scanning tools to look for AI generated or copyrighted material (as mentioned in point 4 of Imaginer's proposal in Discord).

https://discord.com/channels/498741086295031808/535606666708910101/1429839515010207775

https://discord.com/channels/498741086295031808/535606666708910101/1436666376344309791

https://docs.google.com/document/d/1OMh3-_DpjHABMMrXtECzLPNxipQMixx32l6K4zqSoYY/edit?tab=t.0#heading=h.da34uq5e74kw