Adds slash to beginning of image path addresses
What: Adds a slash to the beginning of the 3 previously
comitted images.

Why: Typo and probably not having the slash was the incorrect
path, although some testing didn't help clarify this. Addresses
commit comment on the path topic.

Note: when I went to add in the slash in front of the image paths,
VS Code helpfully suggested the "images" folder and when I selected that,
VS Code suggested the images inside that folder.  So putting the slash
in at the front was/is probably the solution. My thanks for the
commit comments.