Adds Moving Hubs Bot procedure
What: Details process for moving the Hubs Bot to
a non-bound room or a non-bound but bridged room on another server.
Helps HF volunteers remember the recommended steps.

Why: We receive occasional requests to meet in non-bound rooms
or requests from participants that do not want to use a Discord account
but still want to attend a HF meeting. The steps are slightly complex,
but some steps could be forgotten in the stress of the moment. Also,
some steps are included that are not necessary, but kind to do, like
notifying the participant directly by email of the new meeting room link.

Note: This topic was discussed at the HF December 3, 2025 Documentation Session.
Still, a review especially with reference to "rooms", "spaces", 'meetings"
and similar wording would be appreciated because it might not be consistent.

Also, I edited this tone towards HF Vounteers instead of the general public.