Add detail and images to commit instructions
What: Adds a Commit Message Instructions section to policy. Adds images
with highlights emphasizing the sections and the space between
each section. Text and images show the COMMIT_EDITMSG tab and
Terminal window during the act of writing a commit. Slightly modified
Example section title to add that the Example contains the four parts,
Title, What, Why, and Note.

Why: Current Example was a little confusing because it was an example
commit about commit messages. The added images are from a standard
run-of-the-mill commit. Also the added images emphasize following the prompts
that both VS Code shows (to enter and save a commit message) and
to place lines between each section. It is hoped that images will help relay
the importants of having the three minimum sections (Title, What, Why)
and that the space between helps with reading (although I
didn't come out and say it helps reviewers with reading).

Note: Streamlined some vocabulary. Title, What, Why, and Note are
"sections" (not "entries" as I toyed with in my draft).

Not meant as exhaustive instructions as this policy is not the place for that.

The last line is sorta hanging. I didn't know if it would be OK to say
"go and work some more and then push your commits" (plural) so I left
it with the implication that a commit *should* be pushed immediately
once it is ready.

A further commit is coming next with some minor text changes in other
parts of this policy. However, each commit is intended to stand alone.

Comments welcome.