Modifies subject/title commit section instructions
What: Changes "subject/title" phrases to simply "title".
Separates instructions about capitalization and lack of period
to one line. Develops the idea that a title should
answer the prompt "If applied, this commit will..."

Why: Reviewers need accurate and adequate titles in order
to figure out what a commit actually entails. Bland or broad
titles don't help in the process of rooting out bugs because
it is hard to determine what commit to examine from a poor
title. Removes "subject" because the word doesn't seem to be used
much in the Hubs commit world.

Note: These changes were discussed at the October 15, 2025
Documentation Session. It is hoped that this reflects what
was intended (except the subject word thing, that's totally my idea).